web/pyproject.toml
[tool.poetry]
name = "Thunderstore"
version = "0.0.0"
description = "Thunderstore is a mod database and API for downloading Risk of Rain 2 mods"
authors = ["Mythic <mythicmaniac@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.0"
django-environ = "^0.4"
arrow = "^0.17"
Pillow = "^6.2"
psycopg2-binary = "^2.8"
easy-thumbnails = "^2.7"  # This has a monkeypatch and might be a difficult upgrade
markdown = "^3.2"
gunicorn = { version = "^20.0", extras = ["gevent"] }
social-auth-app-django = "^4.0"
social-auth-core = "^3.3"
whitenoise = "^5.0"
django-storages = { version = "^1.9", extras = ["google", "boto3"] }
djangorestframework = "^3.11"
drf-yasg = "^1.17"
django-ipware = "^2.1"
sentry_sdk = "^0.19"
django-redis = "^4.11"
pymdown-extensions = "^8.0"
pyjwt = { version = "^1.7", extras = ["crypto"] }

[tool.poetry.dev-dependencies]
pytest = "^6.2"
pytest-cov = "^2.8"
pytest-django = "^4.1"
factory_boy = "^3.1"
Faker = "^5.0"
pytest-mock = "^3.4"
django-debug-toolbar = "^2.2"
pytest-env = "^0.6.2"
black = ">=20.8b1"
isort = "^5.6.4"
pre-commit = "^2.9.3"
mypy = ">=0.790"
django-stubs = "^1.7.0"
djangorestframework-stubs = "^1.3.0"

flake8 = "^3.8.2"
flake8-2020 = "^1.6.0"
flake8-broken-line = "^0.3.0"
flake8-bugbear = ">=20.11.1"
flake8-builtins = "^1.5.3"
flake8-commas = "^2.0.0"
flake8-comprehensions = "^3.3.1"
flake8-pie = "^0.6.1"
flake8-printf-formatting = "^1.1.0"
flake8-pytest-style = "^1.3.0"
pep8-naming = "^0.11.1"

[tool.black]
target-version = ["py38"]
exclude = '''
(
  /(
      \.git
    | \.mypy_cache
    | \.venv
    | \.vscode
    | \.idea
  )/
)
'''

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
